requirements.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.18.1
numpy==2.4.4
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.48
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
